Do no treat the chem attribute for the <math> tag specially
chem is an attribute so it cannot appear in the end tag and should not
be highlighted bold. It is already covered by .* in the opening tag
like all other attributes.

This fixes a critical bug where the region for <math chem> would never
close and it simplifies the code.